Moving theme classes into their own package
