+++ app/src/main/java/org/thoughtcrime/securesms/ui/Carousel.kt
@@ -37,6 +37,8 @@ import androidx.compose.ui.unit.IntOffset
 import androidx.compose.ui.unit.dp
 import kotlinx.coroutines.launch
 import network.loki.messenger.R
+import org.thoughtcrime.securesms.ui.theme.LocalColors
+import org.thoughtcrime.securesms.ui.theme.pillShape
 import kotlin.math.absoluteValue
 import kotlin.math.sign
