+++ app/src/main/java/org/thoughtcrime/securesms/ui/components/ButtonStyle.kt
@@ -10,9 +10,9 @@ import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.dp
-import org.thoughtcrime.securesms.ui.baseBold
-import org.thoughtcrime.securesms.ui.extraSmall
-import org.thoughtcrime.securesms.ui.extraSmallBold
+import org.thoughtcrime.securesms.ui.theme.baseBold
+import org.thoughtcrime.securesms.ui.theme.extraSmall
+import org.thoughtcrime.securesms.ui.theme.extraSmallBold
 
 interface ButtonStyle {
     @OptIn(ExperimentalMaterialApi::class)
